Isolated rendering for teleport request toast
