+++ frontend/src/components/SocialSidebar/PlayersList.tsx
@@ -27,7 +27,6 @@ export default function PlayersInTownList(): JSX.Element {
   const { ourPlayer } = useTownController();
   const sorted = players.concat([]);
   const toast = useToast();
-  const teleportRequest = useTeleportRequest();
   sorted.sort((p1, p2) =>
     p1.userName.localeCompare(p2.userName, undefined, { numeric: true, sensitivity: 'base' }),
   );
@@ -35,49 +34,6 @@ export default function PlayersInTownList(): JSX.Element {
   const handleTeleportRequest = (player: PlayerController) => {
     townController.emitTeleportRequest(player);
   };
-
-  const handleTeleportAccept = (acceptedRequest: TeleportRequest) => {
-    townController.emitTeleportAccept(acceptedRequest);
-  };
-
-  if (teleportRequest && teleportRequest.to.id === ourPlayer.id) {
-    const { from } = teleportRequest;
-    console.log('should toast');
-    toast({
-      position: 'bottom-left',
-      duration: 10000,
-      render: ({ onClose }) => (
-        <Box color='white' p={3} bg='blue.500'>
-          {from.userName} is trying to teleport to you. Would you like to accept?
-          <HStack>
-            <Button
-              size='xs'
-              color='green'
-              onClick={() => {
-                console.log('accept teleport request');
-                handleTeleportAccept(teleportRequest);
-                onClose();
-              }}>
-              confirm
-            </Button>
-            <Button
-              size='xs'
-              color='red'
-              onClick={() => {
-                console.log('denied teleport request');
-                onClose();
-              }}>
-              deny
-            </Button>
-          </HStack>
-        </Box>
-      ),
-    });
-  } else if (!teleportRequest) {
-    console.log('teleport request undefined');
-  } else {
-    console.log('teleport request does not match');
-  }
 
   return (
     <Box>
@@ -90,33 +46,35 @@ export default function PlayersInTownList(): JSX.Element {
         {sorted.map(player => (
           <ListItem key={player.id}>
             <Button
-              onClick={() =>
-                toast({
-                  position: 'bottom-left',
-                  duration: 10000,
-                  render: ({ onClose }) => (
-                    <Box color='white' p={3} bg='blue.500'>
-                      Would you like to teleport to {player.userName}?
-                      <HStack>
-                        <Button
-                          size='xs'
-                          color='green'
-                          onClick={() => {
-                            console.log('accept teleport confirm');
-                            // handleTeleport(player);
-                            handleTeleportRequest(player);
-                            onClose();
-                          }}>
-                          confirm
-                        </Button>
-                        <Button size='xs' color='red' onClick={onClose}>
-                          deny
-                        </Button>
-                      </HStack>
-                    </Box>
-                  ),
-                })
-              }>
+              onClick={() => {
+                if (ourPlayer.id !== player.id) {
+                  toast({
+                    position: 'bottom-left',
+                    duration: 10000,
+                    render: ({ onClose }) => (
+                      <Box color='white' p={3} bg='blue.500'>
+                        Would you like to teleport to {player.userName}?
+                        <HStack>
+                          <Button
+                            size='xs'
+                            color='green'
+                            onClick={() => {
+                              console.log('accept teleport confirm');
+                              // handleTeleport(player);
+                              handleTeleportRequest(player);
+                              onClose();
+                            }}>
+                            confirm
+                          </Button>
+                          <Button size='xs' color='red' onClick={onClose}>
+                            deny
+                          </Button>
+                        </HStack>
+                      </Box>
+                    ),
+                  });
+                }
+              }}>
               <PlayerName player={player} />
             </Button>
           </ListItem>
